Add files via upload
The survey report presents the findings of an Importance-Performance Analysis (IPA) conducted using IBM SPSS software
The data focuses on Importance and Performance of various dimensions of Mercer University.  
Objective is to find the key dimensions which needs more attention.
The IPA framework will guide us in pinpointing which aspects of the university’s operations are most crucial to students and need improvement.
Providing recommendation based on the insights.